pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sawmill"
version = "0.1.0"
description = "A configurable TUI for post-processing and filtering EDA tool logs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sawmill Contributors"}
]
keywords = ["tui", "log", "filter", "eda", "terminal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "textual>=0.40.0",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
    "pydantic>=2.0.0",
    "rich-click>=1.7.0",
    "pluggy>=1.3.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[project.entry-points."sawmill.plugins"]
vivado = "sawmill.plugins.vivado:VivadoPlugin"

[project.scripts]
sawmill = "sawmill.__main__:cli"

[project.urls]
Homepage = "https://github.com/sawmill/sawmill"
Documentation = "https://github.com/sawmill/sawmill#readme"
Repository = "https://github.com/sawmill/sawmill"

[tool.hatch.build.targets.wheel]
packages = ["sawmill"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
